requirements.txt
flask
flask_wtf
flask-sqlalchemy
python-crontab

